perf tools: Add 'perf-config' command
The perf configuration file contains many variables which can make
the perf command's action more effective.
But looking through state of configuration is difficult and there's no knowing
what kind of other variables except variables in perfconfig.example exist.
So This patch adds 'perf-config' command with '--list' option and a document for it.

    perf config [options]

    display current perf config variables.
    # perf config
    or
    # perf config -l | --list

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>