perf config: Add '--system' and '--global' options to be able to select which config file to be used.
Which config file is used is decided at only perf_config().
That's why a config file to be used can't be selected
among system and global config file when perf-config commands
get or set it.  So add '--system' and '--global' options to select
which config file to be used when getting, setting or etc.

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>