perf config: If failed at collect_config(), finally free a config set after it is done
Because of die() at perf_parse_file() if collect_config() failed,
a config set was freed in collect_config().
But it is natural to free a config set after collect_config() is done.

So, in case of failure, free a config set at perf_config_set__new()
instead of freeing the config set in collect_config().

Cc: Namhyung Kim <namhyung@kernel.org>
Cc: Jiri Olsa <jolsa@redhat.com>
Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Signed-off-by: Taeung Song <treeze.taeung@gmail.com>