perf config: Handle errors from {show_spec, set}_config() in the loop
show_spec_config() and set_config() can be called multiple times
in the loop in cmd_config(). However, the return values from them did not checked.
Even though one of the return values means a error (i.e. -1),
it can be just skipped. It is a bug, so fix it.

Check each return value from show_spec_config() and set_config()
in the loop in cmd_config() and if the return value means a error, handle it.

Cc: Jiri Olsa <jolsa@kernel.org>
Cc: Namhyung Kim <namhyung@kernel.org>
Reported-by: Arnaldo Carvalho de Melo <acme@redhat.com>
Signed-off-by: Taeung Song <treeze.taeung@gmail.com>